+++ samples/core/Samples.sln
@@ -63,6 +63,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "QueryFiltersNavigations", "
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Collations", "Miscellaneous\Collations\Collations.csproj", "{62C86664-49F4-4C59-A2EC-1D70D85149D9}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Async", "Miscellaneous\Async\Async.csproj", "{1DA2B6AD-F71A-4224-92EB-3D0EE6E68BF4}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -467,6 +469,22 @@ Global
 		{62C86664-49F4-4C59-A2EC-1D70D85149D9}.Release|x64.Build.0 = Release|Any CPU
 		{62C86664-49F4-4C59-A2EC-1D70D85149D9}.Release|x86.ActiveCfg = Release|Any CPU
 		{62C86664-49F4-4C59-A2EC-1D70D85149D9}.Release|x86.Build.0 = Release|Any CPU
+		{1DA2B6AD-F71A-4224-92EB-3D0EE6E68BF4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1DA2B6AD-F71A-4224-92EB-3D0EE6E68BF4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1DA2B6AD-F71A-4224-92EB-3D0EE6E68BF4}.Debug|ARM.ActiveCfg = Debug|Any CPU
+		{1DA2B6AD-F71A-4224-92EB-3D0EE6E68BF4}.Debug|ARM.Build.0 = Debug|Any CPU
+		{1DA2B6AD-F71A-4224-92EB-3D0EE6E68BF4}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{1DA2B6AD-F71A-4224-92EB-3D0EE6E68BF4}.Debug|x64.Build.0 = Debug|Any CPU
+		{1DA2B6AD-F71A-4224-92EB-3D0EE6E68BF4}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{1DA2B6AD-F71A-4224-92EB-3D0EE6E68BF4}.Debug|x86.Build.0 = Debug|Any CPU
+		{1DA2B6AD-F71A-4224-92EB-3D0EE6E68BF4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1DA2B6AD-F71A-4224-92EB-3D0EE6E68BF4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1DA2B6AD-F71A-4224-92EB-3D0EE6E68BF4}.Release|ARM.ActiveCfg = Release|Any CPU
+		{1DA2B6AD-F71A-4224-92EB-3D0EE6E68BF4}.Release|ARM.Build.0 = Release|Any CPU
+		{1DA2B6AD-F71A-4224-92EB-3D0EE6E68BF4}.Release|x64.ActiveCfg = Release|Any CPU
+		{1DA2B6AD-F71A-4224-92EB-3D0EE6E68BF4}.Release|x64.Build.0 = Release|Any CPU
+		{1DA2B6AD-F71A-4224-92EB-3D0EE6E68BF4}.Release|x86.ActiveCfg = Release|Any CPU
+		{1DA2B6AD-F71A-4224-92EB-3D0EE6E68BF4}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -490,6 +508,7 @@ Global
 		{E8AD02D7-8AFB-4233-BDAF-F0AEF986F1F3} = {8695C7BE-F9B2-477A-AD7B-C15DC5418F66}
 		{34C237C8-DD12-4C14-9B15-B7F85C218CDB} = {8695C7BE-F9B2-477A-AD7B-C15DC5418F66}
 		{62C86664-49F4-4C59-A2EC-1D70D85149D9} = {85AFD7F1-6943-40FE-B8EC-AA9DBB42CCA6}
+		{1DA2B6AD-F71A-4224-92EB-3D0EE6E68BF4} = {85AFD7F1-6943-40FE-B8EC-AA9DBB42CCA6}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {20C98D35-54EF-46A6-8F3B-1855C1AE4F70}
