--- samples/core/Saving/Async/Blog.cs
@@ -1,12 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace EFSaving.Async
-{
-    public class Blog
-    {
-        public int BlogId { get; set; }
-        public string Url { get; set; }
-
-        public List<Post> Posts { get; set; }
-    }
-}
