--- samples/core/Saving/Async/Post.cs
@@ -1,12 +1,0 @@
-﻿namespace EFSaving.Async
-{
-    public class Post
-    {
-        public int PostId { get; set; }
-        public string Title { get; set; }
-        public string Content { get; set; }
-
-        public int BlogId { get; set; }
-        public Blog Blog { get; set; }
-    }
-}
