+++ samples/core/Saving/Program.cs
@@ -12,7 +12,6 @@
             Transactions.SharingTransaction.Sample.Run();
             Transactions.ExternalDbTransaction.Sample.Run();
             ExplicitValuesGenerateProperties.Sample.Run();
-            Async.Sample.RunAsync().Wait();
             Disconnected.Sample.Run();
         }
     }
